fix: 重複しているファイルを削除
